Move more functions into `docs_utils.py`
