+++ straxen/docs_utils.py
@@ -1,5 +1,7 @@
 from m2r import convert
 
+from .misc import kind_colors
+
 header = """
 Release notes
 ==============
@@ -33,3 +35,41 @@ def convert_release_notes(notes, target, pull_url):
 
     with open(target, "w") as f:
         f.write(header + with_ref)
+
+
+def add_spaces(x):
+    """Add four spaces to every line in x.
+
+    This is needed to make html raw blocks in rst format correctly
+
+    """
+    y = ""
+    if isinstance(x, str):
+        x = x.split("\n")
+    for q in x:
+        y += "    " + q
+    return y
+
+
+def add_deps_to_graph_tree(graph_tree, plugin, data_type, _seen=None):
+    """Recursively add nodes to graph base on plugin.deps."""
+    if _seen is None:
+        _seen = []
+    if data_type in _seen:
+        return graph_tree, _seen
+
+    # Add new one
+    graph_tree.node(
+        data_type,
+        style="filled",
+        href="#" + data_type.replace("_", "-"),
+        fillcolor=kind_colors.get(plugin.data_kind_for(data_type), "grey"),
+    )
+    for dep in plugin.depends_on:
+        graph_tree.edge(data_type, dep)
+
+    # Add any of the lower plugins if we have to
+    for lower_data_type, lower_plugin in plugin.deps.items():
+        graph_tree, _seen = add_deps_to_graph_tree(graph_tree, lower_plugin, lower_data_type, _seen)
+    _seen.append(data_type)
+    return graph_tree, _seen
